# pkg/utils/response.go
package utils

import (
	"encoding/json"
	"log"
	"net/http"
	"time"
)

type APIResponse[T any] struct {
	Success   bool      `json:"success"`
	Data      T         `json:"data,omitempty"`
	Error     *APIError `json:"error,omitempty"`
	Timestamp time.Time `json:"timestamp"`
}

type APIError struct {
	Code    int            `json:"code"`
	Message string         `json:"message"`
	Details map[string]any `json:"details,omitempty"`
}

func WriteResponse[T any](w http.ResponseWriter, status int, data T) {
	response := APIResponse[T]{
		Success:   status >= 200 && status < 300,
		Data:      data,
		Timestamp: time.Now(),
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(response); err != nil {
		log.Printf("failed to encode API response: %v", err)
	}
}

func WriteError(w http.ResponseWriter, status int, message string, details map[string]any) {
	response := APIResponse[any]{
		Success: false,
		Error: &APIError{
			Code:    status,
			Message: message,
			Details: details,
		},
		Timestamp: time.Now(),
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(response); err != nil {
		log.Printf("failed to encode API error response: %v", err)
	}
}

func WriteSuccess(w http.ResponseWriter, data any) {
	WriteResponse(w, http.StatusOK, data)
}

func WriteCreated(w http.ResponseWriter, data any) {
	WriteResponse(w, http.StatusCreated, data)
}

func WriteBadRequest(w http.ResponseWriter, message string) {
	WriteError(w, http.StatusBadRequest, message, nil)
}

func WriteNotFound(w http.ResponseWriter, message string) {
	WriteError(w, http.StatusNotFound, message, nil)
}

func WriteInternalError(w http.ResponseWriter, message string) {
	WriteError(w, http.StatusInternalServerError, message, nil)
}

func WriteUnauthorized(w http.ResponseWriter, message string) {
	WriteError(w, http.StatusUnauthorized, message, nil)
}

func WriteNoContent(w http.ResponseWriter) {
	w.WriteHeader(http.StatusNoContent)
}